requirements.txt
langchain
langchain-community
langchain-anthropic
langchain-openai
langgraph
pydantic>=2.0
python-dotenv
duckduckgo-search
wikipedia
arxiv
beautifulsoup4
requests
tenacity